chore: rename StakingProgramContext.tsx to StakingProgramProvider.tsx
