+++ frontend/context/StakingProgramProvider.tsx
